refactor(web): split status page health/data helpers [OPE-379]
